// expo/src/native/types.gen.ts
export type SetRemoteRequestJson = {
  /**
   * @generated from field: string remote = 1;
   */
  remote?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.SetRemoteResponse
 */
export type SetRemoteResponseJson = {
};

/**
 * @generated from message land.gno.gnonative.v1.GetRemoteRequest
 */
export type GetRemoteRequestJson = {
};

/**
 * @generated from message land.gno.gnonative.v1.GetRemoteResponse
 */
export type GetRemoteResponseJson = {
  /**
   * @generated from field: string remote = 1;
   */
  remote?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.SetChainIDRequest
 */
export type SetChainIDRequestJson = {
  /**
   * @generated from field: string chain_id = 1;
   */
  chainId?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.SetChainIDResponse
 */
export type SetChainIDResponseJson = {
};

/**
 * @generated from message land.gno.gnonative.v1.GetChainIDRequest
 */
export type GetChainIDRequestJson = {
};

/**
 * @generated from message land.gno.gnonative.v1.GetChainIDResponse
 */
export type GetChainIDResponseJson = {
  /**
   * @generated from field: string chain_id = 1;
   */
  chainId?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.SetPasswordRequest
 */
export type SetPasswordRequestJson = {
  /**
   * @generated from field: string password = 1;
   */
  password?: string;

  /**
   * The address of the account to set the password
   *
   * @generated from field: bytes address = 2;
   */
  address?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.SetPasswordResponse
 */
export type SetPasswordResponseJson = {
};

/**
 * @generated from message land.gno.gnonative.v1.RenameKeyRequest
 */
export type RenameKeyRequestJson = {
  /**
   * @generated from field: string old_name = 1;
   */
  oldName?: string;

  /**
   * @generated from field: string new_name = 2;
   */
  newName?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.RenameKeyResponse
 */
export type RenameKeyResponseJson = {
};

/**
 * @generated from message land.gno.gnonative.v1.RotatePasswordRequest
 */
export type RotatePasswordRequestJson = {
  /**
   * @generated from field: string new_password = 1;
   */
  newPassword?: string;

  /**
   * The addresses of the account to rotate the password
   *
   * @generated from field: repeated bytes addresses = 2;
   */
  addresses?: string[];
};

/**
 * @generated from message land.gno.gnonative.v1.RotatePasswordResponse
 */
export type RotatePasswordResponseJson = {
};

/**
 * @generated from message land.gno.gnonative.v1.GenerateRecoveryPhraseRequest
 */
export type GenerateRecoveryPhraseRequestJson = {
};

/**
 * @generated from message land.gno.gnonative.v1.GenerateRecoveryPhraseResponse
 */
export type GenerateRecoveryPhraseResponseJson = {
  /**
   * @generated from field: string phrase = 1;
   */
  phrase?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.KeyInfo
 */
export type KeyInfoJson = {
  /**
   * 0: local, 1: ledger, 2: offline, 3: multi
   *
   * @generated from field: uint32 type = 1;
   */
  type?: number;

  /**
   * @generated from field: string name = 2;
   */
  name?: string;

  /**
   * @generated from field: bytes pub_key = 3;
   */
  pubKey?: string;

  /**
   * @generated from field: bytes address = 4;
   */
  address?: string;
};

/**
 * Coin holds some amount of one currency.
 * A negative amount is invalid.
 *
 * @generated from message land.gno.gnonative.v1.Coin
 */
export type CoinJson = {
  /**
   * Example: "ugnot"
   *
   * @generated from field: string denom = 1;
   */
  denom?: string;

  /**
   * @generated from field: sint64 amount = 2;
   */
  amount?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.BaseAccount
 */
export type BaseAccountJson = {
  /**
   * @generated from field: bytes address = 1;
   */
  address?: string;

  /**
   * @generated from field: repeated land.gno.gnonative.v1.Coin coins = 2;
   */
  coins?: CoinJson[];

  /**
   * @generated from field: bytes pub_key = 3;
   */
  pubKey?: string;

  /**
   * @generated from field: uint64 account_number = 4;
   */
  accountNumber?: string;

  /**
   * @generated from field: uint64 sequence = 5;
   */
  sequence?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.SessionAccount
 */
export type SessionAccountJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.BaseAccount base_account = 1;
   */
  baseAccount?: BaseAccountJson;

  /**
   * @generated from field: bytes master_address = 2;
   */
  masterAddress?: string;

  /**
   * Unix timestamp; 0 = no expiry
   *
   * @generated from field: sint64 expires_at = 3;
   */
  expiresAt?: string;

  /**
   * Nil/empty = no spending allowed (fail-closed, NOT unrestricted)
   *
   * @generated from field: repeated land.gno.gnonative.v1.Coin spend_limit = 4;
   */
  spendLimit?: CoinJson[];

  /**
   * Seconds; 0 = lifetime cap (no reset)
   *
   * @generated from field: sint64 spend_period = 5;
   */
  spendPeriod?: string;

  /**
   * Nil/empty = 0 spent
   *
   * @generated from field: repeated land.gno.gnonative.v1.Coin spend_used = 6;
   */
  spendUsed?: CoinJson[];

  /**
   * Unix timestamp; start of current period
   *
   * @generated from field: sint64 spend_reset = 7;
   */
  spendReset?: string;

  /**
   * @generated from field: repeated string allow_paths = 8;
   */
  allowPaths?: string[];
};

/**
 * @generated from message land.gno.gnonative.v1.ListKeyInfoRequest
 */
export type ListKeyInfoRequestJson = {
};

/**
 * @generated from message land.gno.gnonative.v1.ListKeyInfoResponse
 */
export type ListKeyInfoResponseJson = {
  /**
   * @generated from field: repeated land.gno.gnonative.v1.KeyInfo keys = 1 [json_name = "key_info"];
   */
  key_info?: KeyInfoJson[];
};

/**
 * @generated from message land.gno.gnonative.v1.HasKeyByNameRequest
 */
export type HasKeyByNameRequestJson = {
  /**
   * @generated from field: string name = 1;
   */
  name?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.HasKeyByNameResponse
 */
export type HasKeyByNameResponseJson = {
  /**
   * @generated from field: bool has = 1;
   */
  has?: boolean;
};

/**
 * @generated from message land.gno.gnonative.v1.HasKeyByAddressRequest
 */
export type HasKeyByAddressRequestJson = {
  /**
   * @generated from field: bytes address = 1;
   */
  address?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.HasKeyByAddressResponse
 */
export type HasKeyByAddressResponseJson = {
  /**
   * @generated from field: bool has = 1;
   */
  has?: boolean;
};

/**
 * @generated from message land.gno.gnonative.v1.HasKeyByNameOrAddressRequest
 */
export type HasKeyByNameOrAddressRequestJson = {
  /**
   * @generated from field: string name_or_bech32 = 1;
   */
  nameOrBech32?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.HasKeyByNameOrAddressResponse
 */
export type HasKeyByNameOrAddressResponseJson = {
  /**
   * @generated from field: bool has = 1;
   */
  has?: boolean;
};

/**
 * @generated from message land.gno.gnonative.v1.GetKeyInfoByNameRequest
 */
export type GetKeyInfoByNameRequestJson = {
  /**
   * @generated from field: string name = 1;
   */
  name?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.GetKeyInfoByNameResponse
 */
export type GetKeyInfoByNameResponseJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.KeyInfo key = 1 [json_name = "key_info"];
   */
  key_info?: KeyInfoJson;
};

/**
 * @generated from message land.gno.gnonative.v1.GetKeyInfoByAddressRequest
 */
export type GetKeyInfoByAddressRequestJson = {
  /**
   * @generated from field: bytes address = 1;
   */
  address?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.GetKeyInfoByAddressResponse
 */
export type GetKeyInfoByAddressResponseJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.KeyInfo key = 1 [json_name = "key_info"];
   */
  key_info?: KeyInfoJson;
};

/**
 * @generated from message land.gno.gnonative.v1.GetKeyInfoByNameOrAddressRequest
 */
export type GetKeyInfoByNameOrAddressRequestJson = {
  /**
   * @generated from field: string name_or_bech32 = 1;
   */
  nameOrBech32?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.GetKeyInfoByNameOrAddressResponse
 */
export type GetKeyInfoByNameOrAddressResponseJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.KeyInfo key = 1 [json_name = "key_info"];
   */
  key_info?: KeyInfoJson;
};

/**
 * @generated from message land.gno.gnonative.v1.CreateAccountRequest
 */
export type CreateAccountRequestJson = {
  /**
   * @generated from field: string name_or_bech32 = 1;
   */
  nameOrBech32?: string;

  /**
   * @generated from field: string mnemonic = 2;
   */
  mnemonic?: string;

  /**
   * @generated from field: string bip39_passwd = 3;
   */
  bip39Passwd?: string;

  /**
   * @generated from field: string password = 4;
   */
  password?: string;

  /**
   * @generated from field: uint32 account = 5;
   */
  account?: number;

  /**
   * @generated from field: uint32 index = 6;
   */
  index?: number;
};

/**
 * @generated from message land.gno.gnonative.v1.CreateAccountResponse
 */
export type CreateAccountResponseJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.KeyInfo key = 1 [json_name = "key_info"];
   */
  key_info?: KeyInfoJson;
};

/**
 * @generated from message land.gno.gnonative.v1.CreateLedgerRequest
 */
export type CreateLedgerRequestJson = {
  /**
   * @generated from field: string name = 1;
   */
  name?: string;

  /**
   * Supported algorithm is "secp256k1"
   *
   * @generated from field: string algorithm = 2;
   */
  algorithm?: string;

  /**
   * The human readable part of the address. Example: "g"
   *
   * @generated from field: string hrp = 3;
   */
  hrp?: string;

  /**
   * @generated from field: uint32 account = 4;
   */
  account?: number;

  /**
   * @generated from field: uint32 index = 5;
   */
  index?: number;
};

/**
 * @generated from message land.gno.gnonative.v1.CreateLedgerResponse
 */
export type CreateLedgerResponseJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.KeyInfo key = 1 [json_name = "key_info"];
   */
  key_info?: KeyInfoJson;
};

/**
 * @generated from message land.gno.gnonative.v1.ActivateAccountRequest
 */
export type ActivateAccountRequestJson = {
  /**
   * @generated from field: string name_or_bech32 = 1;
   */
  nameOrBech32?: string;

  /**
   * (Optional) The address of the master account if this is a session account.
   *
   * @generated from field: bytes master = 2;
   */
  master?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.ActivateAccountResponse
 */
export type ActivateAccountResponseJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.KeyInfo key = 1 [json_name = "key_info"];
   */
  key_info?: KeyInfoJson;

  /**
   * True if the password has been set. If false, then call SetPassword.
   *
   * @generated from field: bool has_password = 2;
   */
  hasPassword?: boolean;
};

/**
 * @generated from message land.gno.gnonative.v1.GetActivatedAccountRequest
 */
export type GetActivatedAccountRequestJson = {
  /**
   * @generated from field: bytes address = 1;
   */
  address?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.GetActivatedAccountResponse
 */
export type GetActivatedAccountResponseJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.KeyInfo key = 1 [json_name = "key_info"];
   */
  key_info?: KeyInfoJson;

  /**
   * The Master which was given to ActivateAccount.
   *
   * @generated from field: bytes master = 2;
   */
  master?: string;

  /**
   * True if the password has been set. If false, then call SetPassword.
   *
   * @generated from field: bool has_password = 3;
   */
  hasPassword?: boolean;
};

/**
 * @generated from message land.gno.gnonative.v1.QueryAccountRequest
 */
export type QueryAccountRequestJson = {
  /**
   * @generated from field: bytes address = 1;
   */
  address?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.QueryAccountResponse
 */
export type QueryAccountResponseJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.BaseAccount account_info = 1;
   */
  accountInfo?: BaseAccountJson;
};

/**
 * @generated from message land.gno.gnonative.v1.QuerySessionAccountRequest
 */
export type QuerySessionAccountRequestJson = {
  /**
   * @generated from field: bytes master_address = 1;
   */
  masterAddress?: string;

  /**
   * @generated from field: bytes session_address = 2;
   */
  sessionAddress?: string;
};

/**
 * @generated from message land.gno.gnonative.v1.QuerySessionAccountResponse
 */
export type QuerySessionAccountResponseJson = {
  /**
   * @generated from field: land.gno.gnonative.v1.SessionAccount account_info = 1;
   */
  accountInfo?: SessionAccountJson;
};
// ...
export type DeleteAccountResponseJson = {
};